feat(android): 数据模型契约对齐 + Repository 统一改造
模型层：新增 DashboardVO/ElectronicResourceVO/NegotiationVO/SuggestVO/
SupplierVO；BorrowRecordVO.fineAmount Double→BigDecimal；
BookSimpleVO 新增静态工厂 from(Book,categoryName)；ReservationVO 补 notifyTime；
NegotiationSuggestion.PriceRange Double→BigDecimal。
Repository 层：7 个 Repository 全部接入 ApiCallExecutor 替代裸 execute().body()；
新增 AcquisitionRepository（含 SSE 流式解析）支持 5 个采编端点。
LibraryApi 新增 20+ 端点（采编/管理 Dashboard/KG keyPath/rebuild/超期）。

+++ library-android/app/src/main/java/com/library/android/model/NegotiationSuggestion.java
@@ -2,6 +2,7 @@ package com.library.android.model;
 
 import com.google.gson.annotations.SerializedName;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -34,23 +35,24 @@ public class NegotiationSuggestion {
     public List<String> getKeyTerms() { return keyTerms; }
     public List<String> getRiskWarnings() { return riskWarnings; }
 
+    /** WP2.6：金额字段统一使用 BigDecimal 保证金融精度. */
     public static class PriceRange {
         @SerializedName("floorPrice")
-        private double floorPrice;
+        private BigDecimal floorPrice;
 
         @SerializedName("ceilingPrice")
-        private double ceilingPrice;
+        private BigDecimal ceilingPrice;
 
         @SerializedName("medianPrice")
-        private double medianPrice;
+        private BigDecimal medianPrice;
 
         @SerializedName("suggestedOffer")
-        private double suggestedOffer;
+        private BigDecimal suggestedOffer;
 
-        public double getFloorPrice() { return floorPrice; }
-        public double getCeilingPrice() { return ceilingPrice; }
-        public double getMedianPrice() { return medianPrice; }
-        public double getSuggestedOffer() { return suggestedOffer; }
+        public BigDecimal getFloorPrice() { return floorPrice; }
+        public BigDecimal getCeilingPrice() { return ceilingPrice; }
+        public BigDecimal getMedianPrice() { return medianPrice; }
+        public BigDecimal getSuggestedOffer() { return suggestedOffer; }
     }
 
     public static class Strategy {
